services/feature-engineering/requirements.txt
kafka-python==2.0.2
redis==5.0.1
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
TA-Lib==0.4.28
pandas-ta==0.3.14b0
scipy==1.11.4
scikit-learn==1.3.2
vaderSentiment==3.3.2
transformers==4.36.2
torch==2.1.2
prometheus-client==0.19.0
